Created the pop up modal for the resourceForm with styling and functionality
