+++ src/pages/EditResources/EditResources.jsx
@@ -4,14 +4,8 @@ import "./EditResources.scss";
 import SideNav from "../../components/SideNav/SideNav";
 import MobileHomeButton from "../../components/MobileHomeButton/MobileHomeButton";
 import ResourceForm from "../../components/Forms/ResourceForm/ResourceForm";
-import Button from "../../components/Button/Button";
 
 const EditResources = () => {
-  const [popUp, setPopUp] = useState(false);
-
-  const handlePopUp = () => {
-    setPopUp(!popUp);
-  };
   return (
     <div className="edit-resources">
       <SideNav />
@@ -24,17 +18,6 @@ const EditResources = () => {
         <ResourceForm className="edit-resources__container--form" />
 
         <MobileHomeButton />
-      </div>
-      <div className="display-pop-up">
-        <h1>Edit Resource Complete</h1>
-        <Link to="/home">
-          <Button
-            onClick={handlePopUp}
-            name="Home"
-            image={false}
-            variant="yellowMid"
-          />
-        </Link>
       </div>
     </div>
   );
