+++ src/pages/EditResources/EditResources.scss
@@ -6,21 +6,7 @@
     flex-direction: column;
   }
 
-  .display-pop-up {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    left: 450px;
-    max-width: 50%;
-    background-color: white;
-    color: white;
-    border-radius: 5px;
-    text-align: center;
-    z-index: 9999;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    opacity: 1;
-    transition: opacity 2s ease-in-out;
-  }
+ 
 }
 
 @media screen and (min-width: 720px) {
